refactor project structure (create UI_kit folder
