+++ src/UI_kit/ExpansionPanel/ExpansionPanel.css
